refactor(telemetry): split client lifecycle tests
